requirements.txt
pandas
numpy
scikit-learn
tweepy
praw
telethon
nltk
transformers
matplotlib
seaborn
tensorflow
python-dotenv
